Generalize Dart import resolvers
